Adding user-task relationship

+++ src/tasks/tasks.service.ts
@@ -1,5 +1,6 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
+import { User } from '../auth/user.entity';
 import { CreateTaskDto } from './dto/create-task.dto';
 import { GetTasksFilterDto } from './dto/get-tasks-filter.dto';
 import { TaskStatus } from './task-status.enum';
@@ -15,14 +16,15 @@ export class TasksService {
     this.taskRepository = taskRepository;
   }
 
-  async getTasks(filterDto: GetTasksFilterDto): Promise<Task[]> {
-    return this.taskRepository.getTasks(filterDto);
+  async getTasks(filterDto: GetTasksFilterDto, user: User): Promise<Task[]> {
+    return this.taskRepository.getTasks(filterDto, user);
   }
 
-  async getTaskById(id: number): Promise<Task> {
+  async getTaskById(id: number, user: User): Promise<Task> {
     const task = await this.taskRepository.findOne({
       where: {
         id,
+        userId: user.id
       },
     });
 
@@ -31,27 +33,21 @@ export class TasksService {
     return task;
   }
 
-  async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-    return this.taskRepository.createTask(createTaskDto);
+  async createTask(createTaskDto: CreateTaskDto, user: User): Promise<Task> {
+    return this.taskRepository.createTask(createTaskDto, user);
   }
 
-  async deleteTask(id: number): Promise<void> {
-    const toRemove = await this.taskRepository.delete(id);
+  async deleteTask(id: number, user: User): Promise<void> {
+    const toRemove = await this.taskRepository.delete({ id, userId: user.id });
 
     if (toRemove.affected === 0) throw new NotFoundException();
   }
 
-  async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-    const task = await this.getTaskById(id);
+  async updateTaskStatus(id: number, status: TaskStatus, user: User): Promise<Task> {
+    const task = await this.getTaskById(id, user);
     task.status = status;
     await task.save();
 
     return task;
   }
-
-  // updateTaskStatus(id: string, status: TaskStatus): Task {
-  //   let task = this.getTaskById(id);
-  //   task.status = status;
-  //   return task;
-  // }
 }
